refactor(Checkbox): useMergeRefsを使うinput部分をclient/ActualCheckboxに切り出し'use client'を削除
