+++ packages/smarthr-ui/src/components/Checkbox/client/index.ts
@@ -1,0 +1,1 @@
+export { ActualCheckbox } from './ActualCheckbox'
